Move PatchLightmapColorAssert() to PathSettings.xaml.cs, run on profile save
